requirements.txt
pandas>=1.3.0
requests>=2.25.1
python-dotenv>=0.19.0
tabulate>=0.8.9
openpyxl>=3.0.7
